brain-python/requirements.txt
fastapi
uvicorn[standard]
ollama
openai-whisper
pypdf2
python-multipart
httpx
sentence-transformers
apscheduler
asyncpg
redis
